Refactor `mshio::Data` (#25)
* Refactor post processing data and fix warnings.

* Update python binding.

* Compilation fix.

* Warning and compilation fixes.

* Windows fix.

* Add missing include.

* Wrong include.

+++ src/load_msh_post_process.cpp
@@ -25,8 +25,8 @@ void regroup_nodes_into_blocks(MshSpec& spec)
         for (size_t i = 0; i < block.num_elements_in_block; i++) {
             for (size_t j = 0; j < n; j++) {
                 size_t idx = node_index(block.data[i * (n + 1) + j + 1]);
-                entity_dims[idx] = block.entity_dim;
-                entity_tags[idx] = block.entity_tag;
+                entity_dims[idx] = static_cast<size_t>(block.entity_dim);
+                entity_tags[idx] = static_cast<size_t>(block.entity_tag);
             }
         }
     }
@@ -76,7 +76,7 @@ void regroup_elements_into_blocks(MshSpec& spec)
     std::vector<ElementBlock> element_blocks;
     element_blocks.reserve(elements.num_entity_blocks);
 
-    size_t curr_dim = 0, curr_tag = 0, curr_element_type = 0;
+    int curr_dim = 0, curr_tag = 0, curr_element_type = 0;
     for (auto& block : elements.entity_blocks) {
         if (block.entity_dim != curr_dim || block.entity_tag != curr_tag ||
             block.element_type != curr_element_type) {
